style(src/components/inputs): add example for radio
add example component for radio buttons logic
